python/drug/requirements.txt
asgiref==3.9.1
attrs==25.3.0
certifi==2025.8.3
charset-normalizer==3.4.3
coverage==7.10.6
Django==5.2.5
django-cors-headers==4.7.0
django-environ==0.12.0
django-filter==25.1
django-phonenumber-field==8.1.0
djangorestframework==3.16.1
drf-spectacular==0.28.0
factory_boy==3.3.3
Faker==37.5.3
idna==3.10
inflection==0.5.1
iniconfig==2.1.0
jsonschema==4.25.1
jsonschema-specifications==2025.4.1
packaging==25.0
phonenumbers==9.0.12
pillow==11.3.0
pluggy==1.6.0
psycopg2==2.9.10
pyaes==1.6.1
Pygments==2.19.2
Pyrogram==2.0.106
PySocks==1.7.1
pytest==8.4.1
pytest-cov==6.2.1
pytest-django==4.11.1
PyYAML==6.0.2
referencing==0.36.2
requests==2.32.5
rpds-py==0.27.1
sqlparse==0.5.3
TgCrypto==1.2.5
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
